Remove need for force-unwrap with failable internal initializer
The force unwap in `ActionProtocol.init(input:_:)` is a code smell
warning of a possible race condition between checking the enabled state
of the input property and unwrapping the input value. It's difficult to
guarantee that this is safe in all conditions.

This change modifies the designated initializer to allow the execute
closure to send `ActionError` instead of just `Error`. This gives us the
ability to safely unwrap the input property, and if the value is nil go
ahead and send `ActionError.disabled`.

It's not appropriate to force all consumers to change their code to
return `ActionError`, so this new initialiser has to be an unambiguous
overload of `Action.init(enabledIf:_:)`. Because of trailing closure
syntax, the only way to make the overload unambiguous is to change the
external parameter labels. This change uses the label
`_internalEnabledIf:` to signal to consumers that this is an
implementation detail that shouldn't be called directly.

Once Swift removes the 'Same-type requirement makes generic parameter
non-generic' error, we'll be able to implement `Action.init(input:_:)`
as en extension on `Action` instead of `ActionProtocol`, and the
internal initializer can truly be marked internal at that point.