Add Action initializer for non-Void inputs that also unwraps state
It is valid to conditionally unwrap an optional state property, but also
want to accept input of a different type. This change adds another
convenience initialiser to `Action` that conditionally unwraps, but also
preserves the `input` value passed to `apply()`.

The new overload uses the parameter label `unwrapping` in order to
distinguish it from the version of the initialiser that is generic over
`T`. Unfortunately, even if you use a property of optional values, the
compiler still selects the more-generic version that accepts any `T`
unless a different parameter label is used.