Fix `schedule(after:interval:leeway:)` disposal (#584)
* Fix `schedule(after:interval:leeway:)` disposal

Retain timers in `QueueScheduler` so that they will not be cancelled if the user doesn’t retain the returned `Disposable`.

* Add pull request # and handle

* Use `Atomic` instead of `Lock`

* Separate the tests for repeated action disposal

* Use `Set` instead of `Array` to store timers

* Use `ObjectIdentifier(value).hashValue`

* Weakify `self` in the returned `Disposable`

* Fix `QueueScheduler` tests

* Fix increment in test

Specify `pollInterval` on the asynchronous expectation to avoid false positive test success.

* Conform to timer wrapper to `Hashable` by self ref

`DispatchSourceTimer` doesn’t inherit from `NSObjectProtocol` on Linux. Basing hashing and equality on the wrapper is sufficient because the  same wrapper instance can be used insert and remove from the `Set` of retained timers.