Make advance by TimeInterval in TestScheduler possible
The `currentDate` in `TestScheduler could already be advanced by
`DispatchTimeInterval`. In unit tests the developer might want to
reuse an existing constant, which often is a `Double`/`TimeInterval`.

This commits adds a second method with exactly the same signature as
the existing one; only the type of the argument is different.